% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.adcon}
\alias{process.adcon}
\title{Process ADCON data.}
\usage{
process.adcon(dirFTP, dirAWS, dirUP = NULL, upload = TRUE)
}
\arguments{
\item{dirFTP}{full path to the directory containing the Adcon data.\cr}

\item{dirAWS}{full path to the directory containing the AWS_DATA folder on Adcon server.\cr}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.
Default NULL, must be provided if \code{upload} is \code{TRUE}.\cr}

\item{upload}{logical, if TRUE the data will be uploaded to ADT server.}
}
\description{
Get the data from ADCON database, parse and insert into ADT database.
}
