% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.tahmo}
\alias{process.tahmo}
\title{Process TAHMO data from API.}
\usage{
process.tahmo(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder on ADT server.\cr}
}
\description{
Get the data from TAHMO API, parse and insert into ADT database.
}
