% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.tahmo0}
\alias{process.tahmo0}
\title{Process TAHMO data.}
\usage{
process.tahmo0(dirTAHMO, dirAWS, dirUP = NULL, upload = TRUE)
}
\arguments{
\item{dirTAHMO}{full path to the directory containing the Tahmo data.\cr}

\item{dirAWS}{full path to the directory containing the AWS_DATA folder on Tahmo server.\cr}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.
Default NULL, must be provided if \code{upload} is \code{TRUE}.\cr}

\item{upload}{logical, if TRUE the data will be uploaded to ADT server.}
}
\description{
Get the data from TAHMO database, parse and insert into ADT database.
}
